refactor(sidebar): fold filter into view-options dropdown
Consolidates Status/Repositories filters into the existing view-options
menu next to '+', so there's no layout contention between a filter chip
and the other header controls. An 'active filter' dot on the gear
preserves the affordance that filters are applied.
